fixed page offset and organized the patients-table mocks folder a bit
